Move CanvasLayer to core/util/canvas
