+++ bokehjs/src/lib/models/canvas/canvas.ts
@@ -2,13 +2,12 @@ import {HasProps} from "core/has_props"
 import {DOMView} from "core/dom_view"
 import {logger} from "core/logging"
 import * as p from "core/properties"
-import {div, canvas, append} from "core/dom"
+import {div, append} from "core/dom"
 import {OutputBackend} from "core/enums"
 import {extend} from "core/util/object"
 import {UIEventBus} from "core/ui_events"
 import {BBox} from "core/util/bbox"
-import {Context2d, fixup_ctx} from "core/util/canvas"
-import {SVGRenderingContext2D} from "core/util/svg"
+import {Context2d, CanvasLayer} from "core/util/canvas"
 import {PlotView} from "../plots/plot"
 
 export type FrameBox = [number, number, number, number]
@@ -51,98 +50,6 @@ const style = {
   left: "0",
   width: "100%",
   height: "100%",
-}
-
-export class CanvasLayer {
-  private readonly _canvas: HTMLCanvasElement | SVGSVGElement
-  get canvas(): HTMLCanvasElement {
-    return this._canvas as HTMLCanvasElement
-  }
-
-  private readonly _ctx: CanvasRenderingContext2D | SVGRenderingContext2D
-  get ctx(): Context2d {
-    return this._ctx as Context2d
-  }
-
-  private readonly _el: HTMLElement
-  get el(): HTMLElement {
-    return this._el
-  }
-
-  readonly pixel_ratio: number = 1
-
-  bbox: BBox = new BBox()
-
-  constructor(readonly backend: OutputBackend, readonly hidpi: boolean) {
-    switch (backend) {
-      case "webgl":
-      case "canvas": {
-        this._el = this._canvas = canvas({style})
-        const ctx = this.canvas.getContext('2d')
-        if (ctx == null)
-          throw new Error("unable to obtain 2D rendering context")
-        this._ctx = ctx
-        if (hidpi) {
-          this.pixel_ratio = devicePixelRatio
-        }
-        break
-      }
-      case "svg": {
-        const ctx = new SVGRenderingContext2D()
-        this._ctx = ctx
-        this._canvas = ctx.get_svg()
-        this._el = div({style}, this._canvas)
-        break
-      }
-    }
-
-    fixup_ctx(this._ctx)
-  }
-
-  resize(width: number, height: number): void {
-    this.bbox = new BBox({left: 0, top: 0, width, height})
-
-    const target = this._ctx instanceof SVGRenderingContext2D ? this._ctx : this.canvas
-    target.width = width*this.pixel_ratio
-    target.height = height*this.pixel_ratio
-  }
-
-  prepare(): void {
-    const {ctx, hidpi, pixel_ratio} = this
-    ctx.save()
-    if (hidpi) {
-      ctx.scale(pixel_ratio, pixel_ratio)
-      ctx.translate(0.5, 0.5)
-    }
-    this.clear()
-  }
-
-  clear(): void {
-    const {x, y, width, height} = this.bbox
-    this.ctx.clearRect(x, y, width, height)
-  }
-
-  finish(): void {
-    this.ctx.restore()
-  }
-
-  to_blob(): Promise<Blob> {
-    const {_canvas} = this
-    if (_canvas instanceof HTMLCanvasElement) {
-      if (_canvas.msToBlob != null) {
-        return Promise.resolve(_canvas.msToBlob())
-      } else {
-        return new Promise((resolve, reject) => {
-          _canvas.toBlob((blob) => blob != null ? resolve(blob) : reject(), "image/png")
-        })
-      }
-    } else {
-      const ctx = this._ctx as SVGRenderingContext2D
-      const svg = ctx.get_serialized_svg(true)
-      const blob = new Blob([svg], {type: "image/svg+xml"})
-      return Promise.resolve(blob)
-    }
-  }
 }
 
 export class CanvasView extends DOMView {
